feat(secret-rotation): added db ssl support
