+++ frontend/src/views/SecretRotationPage/components/CreateRotationForm/steps/RotationOutputForm.tsx
@@ -2,40 +2,42 @@ import { Controller, useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { z } from "zod";
 
-import { Button, FormControl, Select, SelectItem } from "@app/components/v2";
+import { Button, FormControl, Input, Select, SelectItem, Spinner } from "@app/components/v2";
 import { useWorkspace } from "@app/context";
 import { useGetProjectSecrets, useGetUserWsKey } from "@app/hooks/api";
 
-const formSchema = z.record(z.string());
+const formSchema = z.object({
+  environment: z.string().trim(),
+  secretPath: z.string().trim().default("/"),
+  interval: z.number().min(1),
+  secrets: z.record(z.string())
+});
 
 export type TFormSchema = z.infer<typeof formSchema>;
 type Props = {
-  environment: string;
-  secretPath: string;
   outputSchema: Record<string, unknown>;
   onSubmit: (data: TFormSchema) => void;
   onCancel: () => void;
 };
 
-export const RotationOutputForm = ({
-  onSubmit,
-  onCancel,
-  environment,
-  secretPath,
-  outputSchema = {}
-}: Props) => {
+export const RotationOutputForm = ({ onSubmit, onCancel, outputSchema = {} }: Props) => {
   const { currentWorkspace } = useWorkspace();
+  const environments = currentWorkspace?.environments || [];
   const workspaceId = currentWorkspace?._id || "";
   const {
     control,
     handleSubmit,
+    watch,
     formState: { isSubmitting }
   } = useForm<TFormSchema>({
     resolver: zodResolver(formSchema)
   });
 
+  const environment = watch("environment", environments?.[0]?.slug);
+  const secretPath = watch("secretPath");
+
   const { data: userWsKey } = useGetUserWsKey(workspaceId);
-  const { data: secrets } = useGetProjectSecrets({
+  const { data: secrets, isLoading: isSecretsLoading } = useGetProjectSecrets({
     workspaceId,
     environment,
     secretPath,
@@ -44,11 +46,65 @@ export const RotationOutputForm = ({
 
   return (
     <form onSubmit={handleSubmit(onSubmit)}>
+      <Controller
+        control={control}
+        name="environment"
+        defaultValue={environments?.[0]?.slug}
+        render={({ field: { value, onChange } }) => (
+          <FormControl label="Environment">
+            <Select
+              value={value}
+              onValueChange={(val) => onChange(val)}
+              className="w-full border border-mineshaft-500"
+              defaultValue={environments?.[0]?.slug}
+              position="popper"
+            >
+              {environments.map((sourceEnvironment) => (
+                <SelectItem
+                  value={sourceEnvironment.slug}
+                  key={`source-environment-${sourceEnvironment.slug}`}
+                >
+                  {sourceEnvironment.name}
+                </SelectItem>
+              ))}
+            </Select>
+          </FormControl>
+        )}
+      />
+      <Controller
+        control={control}
+        name="secretPath"
+        defaultValue="/"
+        render={({ field }) => (
+          <FormControl className="capitalize" label="Secret path">
+            <Input {...field} />
+          </FormControl>
+        )}
+      />
+      <Controller
+        control={control}
+        name="interval"
+        defaultValue={15}
+        render={({ field }) => (
+          <FormControl className="capitalize" label="Rotation Interval (Days)">
+            <Input
+              {...field}
+              min={1}
+              type="number"
+              onChange={(evt) => field.onChange(parseInt(evt.target.value, 10))}
+            />
+          </FormControl>
+        )}
+      />
+      <div className="flex flex-col mt-4 pt-4 mb-2 border-t border-bunker-300/30">
+        <div>Mapping</div>
+        <div className="text-bunker-300 text-sm">Select keys for rotated value to get saved</div>
+      </div>
       {Object.keys(outputSchema).map((outputName) => (
         <Controller
           key={`provider-output-${outputName}`}
           control={control}
-          name={outputName}
+          name={`secrets.${outputName}`}
           render={({ field: { value, onChange } }) => (
             <FormControl className="uppercase" label={outputName.replaceAll("_", " ")} isRequired>
               <Select
@@ -57,11 +113,22 @@ export const RotationOutputForm = ({
                 className="w-full border border-mineshaft-500"
                 position="popper"
               >
-                {secrets?.map(({ key, _id }) => (
-                  <SelectItem value={_id} key={_id}>
-                    {key}
+                {!isSecretsLoading &&
+                  secrets?.map(({ key, _id }) => (
+                    <SelectItem value={_id} key={_id}>
+                      {key}
+                    </SelectItem>
+                  ))}
+                {isSecretsLoading && (
+                  <SelectItem value="Loading" isDisabled>
+                    <Spinner size="xs" />
                   </SelectItem>
-                ))}
+                )}
+                {!isSecretsLoading && secrets?.length === 0 && (
+                  <SelectItem value="Empty" isDisabled>
+                    No secrets found
+                  </SelectItem>
+                )}
               </Select>
             </FormControl>
           )}
